+++ monitor/PromiseMonitor.js
@@ -5,13 +5,19 @@
 (function(define) { 'use strict';
 define(function(require) {
 
+	var defaultStackJumpSeparator = 'from execution context:';
+	var defaultStackFilter = /(node|module|timers)\.js:|when(\/(lib|monitor|es6-shim)\/|\.js)|(new\sPromise)\b|(\b(PromiseMonitor|ConsoleReporter|Scheduler|RunHandlerTask|ProgressTask|Promise|.*Handler)\.[\w_]\w\w+\b)|\b(tryCatch\w+|getHandler\w*)\b/i;
+
 	var setTimer = require('../lib/timer').set;
+	var error = require('./error');
 
 	function PromiseMonitor(reporter) {
-		this.key = 0;
 		this.traces = {};
 		this.traceTask = 0;
 		this.logDelay = 100;
+		this.stackFilter = defaultStackFilter;
+		this.stackJumpSeparator = defaultStackJumpSeparator;
+		this.filterDuplicateFrames = true;
 
 		this._reporter = reporter;
 
@@ -21,29 +27,32 @@ define(function(require) {
 		};
 	}
 
-	PromiseMonitor.prototype.captureStack = function() {
-		return new Error();
+	PromiseMonitor.prototype.captureStack = function(host, at) {
+		return error.captureStack(host, at);
 	};
 
-	PromiseMonitor.prototype.startTrace = function(trace) {
-		var key = this.key++;
-		this.traces[key] = [trace];
-		return key;
+	PromiseMonitor.prototype.addTrace = function(handler, extraContext) {
+		this.traces[handler.id] = {
+			error: handler.value,
+			context: handler.context,
+			extraContext: extraContext
+		};
+		this.logTraces();
 	};
 
-	PromiseMonitor.prototype.updateTrace = function(key, trace) {
-		var t = this.traces[key];
-		if(typeof t !== 'undefined') {
-			t.push(trace);
+	PromiseMonitor.prototype.removeTrace = function(handler) {
+		if(handler.id in this.traces) {
+			delete this.traces[handler.id];
 			this.logTraces();
 		}
 	};
 
-	PromiseMonitor.prototype.removeTrace = function(key) {
-		if(key in this.traces) {
-			delete this.traces[key];
-			this.logTraces();
-		}
+	PromiseMonitor.prototype.fatal = function(handler) {
+		var err = new Error();
+		err.stack = this._createLongTrace(handler.value, handler.context).join('\n');
+		setTimer(function() {
+			throw err;
+		}, 0);
 	};
 
 	PromiseMonitor.prototype.logTraces = function() {
@@ -54,8 +63,87 @@ define(function(require) {
 
 	PromiseMonitor.prototype._logTraces = function() {
 		this.traceTask = void 0;
-		this._reporter.log(this.traces);
+		this._reporter.log(this.formatTraces(this.traces));
 	};
+
+
+	PromiseMonitor.prototype.formatTraces = function(traces) {
+		var keys = Object.keys(traces);
+		var formatted = [];
+		var longTrace, t, i;
+
+		for(i=0; i<keys.length; ++i) {
+			t = traces[keys[i]];
+			longTrace = this._createLongTrace(t.error, t.context, t.extraContext);
+			formatted.push(longTrace);
+		}
+
+		return formatted;
+	};
+
+	PromiseMonitor.prototype._createLongTrace = function(e, context, extraContext) {
+		var trace = error.parse(e) || [];
+		trace = filterFrames(this.stackFilter, trace, 0);
+		this._appendContext(trace, context);
+		this._appendContext(trace, extraContext);
+		return this.filterDuplicateFrames ? this._removeDuplicates(trace) : trace;
+	};
+
+	PromiseMonitor.prototype._removeDuplicates = function(trace) {
+		var seen = {};
+		var sep = this.stackJumpSeparator;
+		var count = 0;
+		return trace.reduceRight(function(deduped, line, i) {
+			if(i === 0) {
+				deduped.unshift(line);
+			} else if(line === sep) {
+				if(count > 0) {
+					deduped.unshift(line);
+					count = 0;
+				}
+			} else if(!seen[line]) {
+				seen[line] = true;
+				deduped.unshift(line);
+				++count;
+			}
+			return deduped;
+		}, []);
+	};
+
+	PromiseMonitor.prototype._appendContext = function(trace, context) {
+		trace.push.apply(trace, this._createTrace(context));
+	};
+
+	PromiseMonitor.prototype._createTrace = function(traceChain) {
+		var trace = [];
+		var stack;
+
+		while(traceChain) {
+			stack = error.parse(traceChain);
+
+			if (stack) {
+				stack = filterFrames(this.stackFilter, stack);
+				appendStack(trace, stack, this.stackJumpSeparator);
+			}
+
+			traceChain = traceChain.parent;
+		}
+
+		return trace;
+	};
+
+	function appendStack(trace, stack, separator) {
+		if (stack.length > 1) {
+			stack[0] = separator;
+			trace.push.apply(trace, stack);
+		}
+	}
+
+	function filterFrames(stackFilter, stack) {
+		return stack.filter(function(frame) {
+			return !stackFilter.test(frame);
+		});
+	}
 
 	return PromiseMonitor;
 });