+++ test/monitor/deep-chain.js
@@ -12,16 +12,7 @@
 	define(function(require) {
 
 		require('../../monitor/console');
-
 		var Promise = require('../../when').Promise;
-
-		Promise.resolve().then(function outer() {
-			return Promise.resolve().then(function inner() {
-				return Promise.resolve().then(function evenMoreInner() {
-					foo()
-				});
-			});
-		});
 
 		function f1() {
 			return Promise.resolve(123);
@@ -45,7 +36,7 @@
 		// and this will be logged as well.
 		setTimeout(function() {
 			console.log('*** handling rejection ***');
-			p.otherwise(ok);
+			p.catch(ok);
 		}, 1337);
 
 		function ok(x) {