+++ test/timeout-test.js
@@ -1,6 +1,6 @@
 (function(buster, define) {
 
-var assert, fail, sentinel;
+var assert, fail, sentinel, undef;
 
 assert = buster.assert;
 fail = buster.assertions.fail;
@@ -59,6 +59,68 @@ define('when/timeout-test', function (require) {
 			).ensure(done);
 
 			d.notify(sentinel);
+		},
+
+		'promise.timeout': {
+
+			'should reject after timeout': function(done) {
+				when.defer().promise.timeout(10).then(
+					fail,
+					function(e) {
+						assert(e instanceof Error);
+					}
+				).ensure(done);
+			},
+
+			'should reject after timeout with the provided reason': function(done) {
+				when.defer().promise.timeout(10, sentinel).then(
+					fail,
+					function(reason) {
+						assert.same(reason, sentinel);
+					}
+				).ensure(done);
+			},
+
+			'should reject after timeout with the provided reason, even if undefined': function(done) {
+				when.defer().promise.timeout(10, undef).then(
+					fail,
+					function(reason) {
+						assert.same(reason, undef);
+					}
+				).ensure(done);
+			},
+
+			'should not timeout when rejected before timeout': function(done) {
+				when.reject(sentinel).timeout(10).then(
+					fail,
+					function(val) {
+						assert.same(val, sentinel);
+					}
+				).ensure(done);
+			},
+
+			'should not timeout when forcibly resolved before timeout': function(done) {
+				when.resolve(sentinel).timeout(10).then(
+					function(val) {
+						assert.same(val, sentinel);
+					},
+					fail
+				).ensure(done);
+			},
+
+			'should propagate progress': function(done) {
+				var d = when.defer();
+
+				d.promise.timeout(10).then(null, null,
+					function(val) {
+						assert.same(val, sentinel);
+						d.resolve();
+					}
+				).ensure(done);
+
+				d.notify(sentinel);
+			}
+
 		}
 
 	});